backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.27.0
azure-identity==1.13.0
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dataclasses-json==0.5.7
dnspython==2.3.0
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.192
langchainplus-sdk==0.0.4
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
microsoft-kiota-abstractions==0.5.1
microsoft-kiota-authentication-azure==0.2.0
microsoft-kiota-http==0.4.3
microsoft-kiota-serialization-json==0.3.4
microsoft-kiota-serialization-text==0.2.0
msal==1.22.0
msal-extensions==1.0.0
msgraph-core==1.0.0a4
msgraph-sdk==1.0.0a12
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
pinecone-client==2.2.2
platformdirs==3.5.1
portalocker==2.7.0
-e git+https://github.com/edwardmfho/promptimize.git@b77e30c2d06ef0d52ed6ff6a2d67e514fc787a38#egg=promptimize
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
PyJWT==2.7.0
pymongo==3.6.0
python-box==7.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
tabulate==0.9.0
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
Werkzeug==2.3.6
yarl==1.9.2
